+++ modules/acid/CHANGELOG.md
@@ -8,3 +8,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Added a new `anyToHex` API to the `Transformations` module. This uses a canvas to convert any value to a hex colour #TINY-7480
+
+### Changed
+- `HexColour.fromString` will now normalize the hex value to strip the leading `#` if present and uppercase the values #TINY-7480